[fix] 정적 페이지의 html파일을 불러오지 못하여 pre-render가 되지 않는 오류 수정 - useSearchParams hook을 사용하는 컴포넌트에 suspence를 적용하지 않아 생긴 문제
