+++ src/app/page.tsx
@@ -1,24 +1,13 @@
-'use client';
-
 import SideNavBar from '@/components/bars/SideNavBar';
+import MoveScroll from '@/components/etc/MoveScroll';
 import CareerSection from '@/components/sections/careers/CareerSection';
 import MainSection from '@/components/sections/main/MainSection';
 import ProjectSecton from '@/components/sections/personal-projects/ProjectSection';
 import SkillSection from '@/components/sections/skills/SkillSection';
 import TroubleShootingSection from '@/components/sections/trouble-shooting/TroubleShootingSection';
-import { useSearchParams } from 'next/navigation';
-import { useEffect } from 'react';
+import { Suspense } from 'react';
 
 export default function Home() {
-  const searchParams = useSearchParams();
-  const step = searchParams.get('step');
-
-  useEffect(() => {
-    if (step === null) return window.scrollTo(0, 0);
-
-    document.getElementById(step)?.scrollIntoView({ block: 'start', behavior: 'smooth' });
-  }, [step]);
-
   return (
     <section className="h-full relative">
       <MainSection id="mainSection" />
@@ -27,6 +16,9 @@ export default function Home() {
         <TroubleShootingSection id="troubleShootingSection" />
         <CareerSection id="careerSection" />
         <ProjectSecton id="projectSection" />
+        <Suspense>
+          <MoveScroll />
+        </Suspense>
         <SideNavBar />
       </div>
     </section>
