+++ src/app/projects/siklog/page.tsx
@@ -15,8 +15,9 @@ import {
   projectName,
   troubleShootings,
 } from './constant/index';
-import ScrollControlSection from '@/components/sections/ScrollControlSection';
 import ProjectTroubleShootingArticle from '@/components/articles/project/ProjectTroubleShootingArticle';
+import { Suspense } from 'react';
+import MoveScroll from '@/components/etc/MoveScroll';
 
 export const metadata = {
   title: projectName,
@@ -26,7 +27,7 @@ export const metadata = {
 
 export default function SiklogPage() {
   return (
-    <ScrollControlSection>
+    <section>
       <FadeIn delay={0.3}>
         <ProjectIntroductArticle
           content={`개인 블로그를 가지고 싶어 Next.js App Route 방식을 이용하여 만든 블로그입니다.\n\nNext.js의 API Route와 MongoDB를 이용하여 데이터 관리를 하였으며, AWS를 이용하여 호스팅 및 배포 자동화를 진행하였으나 나 비용 문제로 인해 호스팅 서비스를 Vercel로 변경하였습니다.\n\n블로그라는 특성상 다른 사용자들에게 노출이 되어야 하기 때문에 어떻게 하면 검색엔진이 읽기 쉬운 블로그를 만들 수 있을지를 고민을 하였습니다.\n\nSEO에 대해 고민하고 적용한 결과는 아래와 같습니다.`}
@@ -52,6 +53,9 @@ export default function SiklogPage() {
       <BottomToTopFadeIn delay={0}>
         <ProjectPreviewArticle title="📺 미리 보기" previews={previews} />
       </BottomToTopFadeIn>
-    </ScrollControlSection>
+      <Suspense>
+        <MoveScroll />
+      </Suspense>
+    </section>
   );
 }
