+++ src/app/projects/wegnal/page.tsx
@@ -1,5 +1,4 @@
 import FadeIn from '@/components/gasp/FadeIn';
-import Underline from '@/components/gasp/UnderLine';
 import ProjectNumberOfPeopleArticle from '@/components/articles/project/ProjectNumberOfPeopleArticle';
 import ProjectRoleArticle from '@/components/articles/project/ProjectRoleArticle';
 import ProjectUseSkillArticle from '@/components/articles/project/ProjectUseSkillArticle';
@@ -12,9 +11,10 @@ import {
   projectName,
   troubleShootings,
 } from './constant/index';
-import ScrollControlSection from '@/components/sections/ScrollControlSection';
 import ProjectIntroductArticle from '@/components/articles/project/ProjectIntroductArticle';
 import ProjectTroubleShootingArticle from '@/components/articles/project/ProjectTroubleShootingArticle';
+import { Suspense } from 'react';
+import MoveScroll from '@/components/etc/MoveScroll';
 
 export const metadata = {
   title: projectName,
@@ -23,7 +23,7 @@ export const metadata = {
 
 export default function WegnalPage() {
   return (
-    <ScrollControlSection>
+    <section>
       <FadeIn delay={0.3}>
         <ProjectIntroductArticle content={introduction} projectName={projectName} />
       </FadeIn>
@@ -42,6 +42,9 @@ export default function WegnalPage() {
       <FadeIn delay={0.3}>
         <ProjectNumberOfPeopleArticle content="프론트엔드 1명, 백엔드 1명" />
       </FadeIn>
-    </ScrollControlSection>
+      <Suspense>
+        <MoveScroll />
+      </Suspense>
+    </section>
   );
 }
