+++ src/app/projects/withwe-homepage/page.tsx
@@ -14,8 +14,9 @@ import {
   introduction,
   troubleShootings,
 } from './constant/index';
-import ScrollControlSection from '@/components/sections/ScrollControlSection';
 import ProjectTroubleShootingArticle from '@/components/articles/project/ProjectTroubleShootingArticle';
+import { Suspense } from 'react';
+import MoveScroll from '@/components/etc/MoveScroll';
 
 export const metadata = {
   title: projectName,
@@ -24,7 +25,7 @@ export const metadata = {
 
 export default function WithweHomePage() {
   return (
-    <ScrollControlSection>
+    <section>
       <FadeIn delay={0.3}>
         <ProjectIntroductArticle content={introduction} projectName={projectName} />
       </FadeIn>
@@ -46,6 +47,9 @@ export default function WithweHomePage() {
       <FadeIn delay={0.3}>
         <ProjectNumberOfPeopleArticle content="프론트엔드 1명" />
       </FadeIn>
-    </ScrollControlSection>
+      <Suspense>
+        <MoveScroll />
+      </Suspense>
+    </section>
   );
 }
