--- src/components/sections/ScrollControlSection.tsx
@@ -1,23 +1,0 @@
-'use client';
-
-import { useSearchParams } from 'next/navigation';
-import React, { useEffect } from 'react';
-
-type Props = {
-  children: React.ReactNode;
-};
-
-export default function ScrollControlSection({ children }: Props) {
-  const searchParams = useSearchParams();
-  const step = searchParams.get('step');
-
-  useEffect(() => {
-    if (step === null) return window.scrollTo(0, 0);
-
-    document
-      .getElementById('troubleShootingStep')
-      ?.scrollIntoView({ block: 'start', behavior: 'smooth' });
-  }, [step]);
-
-  return <section className="w-full">{children}</section>;
-}
